Moved record related functions into a separate file.
